add piwik code to the website
It could be very helpful to see some metrics for our site to rethink our
public relations work. I know that we don't want to store more data than
needed but the piwik installation used is configured with the maximum
data protection options so for example if Do not track ist activated we
wouldt track you.